requirements.txt
Flask==2.1.3
line-bot-sdk==3.11.0
openai==1.25.2
gunicorn==22.0.0
